refactor(prober): use upstream kubelet prober with probe-type rewriting
Remove custom ProberImpl in favor of the kubelet's built-in prober.
Rewrite HTTP/TCP/GRPC probes to exec probes (wget/nc) during pod
admission so the upstream prober routes everything through ExecSync.
A sentinel prefix (__sandbox__) signals CRID to exec in the pod sandbox
instead of the app container, ensuring network tools are available.
